simplified utils and move environment to app since authorization is different for aap and idp (no rest call)
