+++ endpoints/authorizations/authorizations.go
@@ -5,15 +5,15 @@ import (
   "github.com/sirupsen/logrus"
   "github.com/gin-gonic/gin"
 
+  "github.com/charmixer/aap/app"
   "github.com/charmixer/aap/client"
-  "github.com/charmixer/aap/environment"
   "github.com/charmixer/aap/gateway/aap"
 )
 
-func GetAuthorizations(env *environment.State) gin.HandlerFunc {
+func GetAuthorizations(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
 
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "GetAuthorizations",
     })
@@ -90,10 +90,10 @@ func GetAuthorizations(env *environment.State) gin.HandlerFunc {
   return gin.HandlerFunc(fn)
 }
 
-func PostAuthorizations(env *environment.State) gin.HandlerFunc {
+func PostAuthorizations(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
 
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "PostAuthorizations",
     })
@@ -171,9 +171,9 @@ func PostAuthorizations(env *environment.State) gin.HandlerFunc {
   return gin.HandlerFunc(fn)
 }
 
-func PutAuthorizations(env *environment.State) gin.HandlerFunc {
+func PutAuthorizations(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "PutAuthorizations",
     })
