+++ endpoints/publishes/publishes.go
@@ -5,7 +5,7 @@ import (
   "github.com/sirupsen/logrus"
   "github.com/gin-gonic/gin"
 
-  "github.com/charmixer/aap/environment"
+  "github.com/charmixer/aap/app"
   "github.com/charmixer/aap/gateway/aap"
   "github.com/charmixer/aap/client"
 
@@ -14,9 +14,9 @@ import (
   "fmt"
 )
 
-func PostPublishes(env *environment.State) gin.HandlerFunc {
+func PostPublishes(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "PostPublishes",
     })
@@ -94,9 +94,9 @@ func PostPublishes(env *environment.State) gin.HandlerFunc {
   return gin.HandlerFunc(fn)
 }
 
-func DeletePublishes(env *environment.State) gin.HandlerFunc {
+func DeletePublishes(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "DeletePublishes",
     })
@@ -108,9 +108,9 @@ func DeletePublishes(env *environment.State) gin.HandlerFunc {
   return gin.HandlerFunc(fn)
 }
 
-func GetPublishes(env *environment.State) gin.HandlerFunc {
+func GetPublishes(env *app.Environment) gin.HandlerFunc {
   fn := func(c *gin.Context) {
-    log := c.MustGet(environment.LogKey).(*logrus.Entry)
+    log := c.MustGet(env.Constants.LogKey).(*logrus.Entry)
     log = log.WithFields(logrus.Fields{
       "func": "GetPublishes",
     })
