+++ utils/utils.go
@@ -6,36 +6,8 @@ import (
   "net"
   "net/http"
   "crypto/rand"
-  "encoding/base64"
-  "time"
-  "fmt"
-  "golang.org/x/oauth2"
-  "golang.org/x/oauth2/clientcredentials"
-  "github.com/sirupsen/logrus"
-  "github.com/gin-gonic/gin"
-  "github.com/gofrs/uuid"
-  hydra "github.com/charmixer/hydra/client"
-
-  "github.com/charmixer/aap/environment"
-  "github.com/charmixer/aap/config"
-  "github.com/charmixer/aap/gateway/aap"
+  "encoding/base64"  
 )
-
-const ERROR_INVALID_ACCESS_TOKEN = 1
-const ERROR_MISSING_BEARER_TOKEN = 2
-const ERROR_MISSING_REQUIRED_SCOPES = 3
-
-type JsonError struct {
-  ErrorCode int `json:"error_code" binding:"required"`
-  Error     string `json:"error" binding:"required"`
-}
-
-type AuthorizationConfig struct {
-  LogKey             string
-  AccessTokenKey     string
-  HydraConfig        *clientcredentials.Config
-  HydraIntrospectUrl string
-}
 
 func GenerateRandomBytes(n int) ([]byte, error) {
   b := make([]byte, n)
@@ -151,286 +123,6 @@ func detectForwardedForIpAndPort(r *http.Request) (string, string) {
         }
     }
     return "", ""
-}
-
-func ProcessMethodOverride(r *gin.Engine) gin.HandlerFunc {
-  return func(c *gin.Context) {
-
-    // Only need to check POST method
-    if c.Request.Method != "POST" {
-      return
-    }
-
-    method := c.Request.Header.Get("X-HTTP-Method-Override")
-    method = strings.ToLower(method)
-    method = strings.TrimSpace(method)
-
-    // Require using method override
-    if method == "" {
-      c.JSON(http.StatusBadRequest, gin.H{"error": "Missing or empty X-HTTP-Method-Override header"})
-      c.Abort()
-      return
-    }
-
-
-    if method == "post" {
-      // if HandleContext is called you will make an infinite loop
-      //c.Next()
-      return
-    }
-
-    if method == "get" {
-      c.Request.Method = "GET"
-      r.HandleContext(c)
-      c.Abort()
-      return
-    }
-
-    if method == "put" {
-      c.Request.Method = "PUT"
-      r.HandleContext(c)
-      c.Abort()
-      return
-    }
-
-    if method == "delete" {
-      c.Request.Method = "DELETE"
-      r.HandleContext(c)
-      c.Abort()
-      return
-    }
-
-    c.JSON(http.StatusBadRequest, gin.H{"error": "Unsupported method"})
-    c.Abort()
-    return
-  }
-}
-
-
-func RequestLogger(logKey string, requestIdKey string, log *logrus.Logger, appFields logrus.Fields) gin.HandlerFunc {
-  fn := func(c *gin.Context) {
-
-    // Start timer
-    start := time.Now()
-    path := c.Request.URL.Path
-    raw := c.Request.URL.RawQuery
-
-    var requestId string = c.MustGet(requestIdKey).(string)
-    requestLog := log.WithFields(appFields).WithFields(logrus.Fields{
-      "request.id": requestId,
-    })
-    c.Set(logKey, requestLog)
-
-    c.Next()
-
-    // Stop timer
-    stop := time.Now()
-    latency := stop.Sub(start)
-
-    ipData, err := GetRequestIpData(c.Request)
-    if err != nil {
-      log.WithFields(appFields).WithFields(logrus.Fields{
-        "func": "RequestLogger",
-      }).Debug(err.Error())
-    }
-
-    forwardedForIpData, err := GetForwardedForIpData(c.Request)
-    if err != nil {
-      log.WithFields(appFields).WithFields(logrus.Fields{
-        "func": "RequestLogger",
-      }).Debug(err.Error())
-    }
-
-    method := c.Request.Method
-    statusCode := c.Writer.Status()
-    errorMessage := c.Errors.ByType(gin.ErrorTypePrivate).String()
-
-    bodySize := c.Writer.Size()
-
-    var fullpath string = path
-    if raw != "" {
-      fullpath = path + "?" + raw
-    }
-
-    log.WithFields(appFields).WithFields(logrus.Fields{
-      "latency": latency,
-      "forwarded_for.ip": forwardedForIpData.Ip,
-      "forwarded_for.port": forwardedForIpData.Port,
-      "ip": ipData.Ip,
-      "port": ipData.Port,
-      "method": method,
-      "status": statusCode,
-      "error": errorMessage,
-      "body_size": bodySize,
-      "path": fullpath,
-      "request.id": requestId,
-    }).Info("")
-  }
-  return gin.HandlerFunc(fn)
-}
-
-func AuthenticationRequired(logKey string, accessTokenKey string) gin.HandlerFunc {
-  fn := func(c *gin.Context) {
-
-    log := c.MustGet(logKey).(*logrus.Entry)
-    log = log.WithFields(logrus.Fields{
-      "func": "AuthenticationRequired",
-    })
-
-    log = log.WithFields(logrus.Fields{"authorization": "bearer"})
-    log.Debug("Looking for access token")
-    var token *oauth2.Token
-    auth := c.Request.Header.Get("Authorization")
-    split := strings.SplitN(auth, " ", 2)
-    if len(split) == 2 || strings.EqualFold(split[0], "bearer") {
-
-      token = &oauth2.Token{
-        AccessToken: split[1],
-        TokenType: split[0],
-      }
-
-      // See #2 of QTNA
-      // https://godoc.org/golang.org/x/oauth2#Token.Valid
-      if token.Valid() == true {
-
-        // See #5 of QTNA
-        log.WithFields(logrus.Fields{"fixme": 1, "qtna": 5}).Debug("Missing check against token-revoked-list to check if token is revoked")
-
-        log.Debug("Authenticated")
-        c.Set(accessTokenKey, token)
-        c.Next() // Authentication successful, continue.
-        return;
-      }
-
-      // Deny by default
-      c.JSON(http.StatusUnauthorized, JsonError{ErrorCode: ERROR_INVALID_ACCESS_TOKEN, Error: "Invalid access token."})
-      c.Abort()
-      return
-    }
-
-    // Deny by default
-    c.JSON(http.StatusUnauthorized, JsonError{ErrorCode: ERROR_MISSING_BEARER_TOKEN, Error: "Authorization: Bearer <token> not found in request"})
-    c.Abort()
-  }
-  return gin.HandlerFunc(fn)
-}
-
-func AuthorizationRequired(env *environment.State, aconf AuthorizationConfig, requiredScopes ...string) gin.HandlerFunc {
-  fn := func(c *gin.Context) {
-
-
-    log := c.MustGet(aconf.LogKey).(*logrus.Entry)
-    log = log.WithFields(logrus.Fields{"func": "AuthorizationRequired"})
-
-    // This is required to be here but should be garantueed by the authenticationRequired function.
-    t, accessTokenExists := c.Get(aconf.AccessTokenKey)
-    if accessTokenExists == false {
-      c.AbortWithStatusJSON(http.StatusForbidden, JsonError{ErrorCode: ERROR_MISSING_BEARER_TOKEN, Error: "No access token found. Hint: Is bearer token missing?"})
-      return
-    }
-    var accessToken *oauth2.Token = t.(*oauth2.Token)
-
-    strRequiredScopes := strings.Join(requiredScopes, " ")
-    log.WithFields(logrus.Fields{"scope": strRequiredScopes}).Debug("Checking required scopes");
-
-    // See #3 of QTNA
-    // log.WithFields(logrus.Fields{"fixme": 1, "qtna": 3}).Debug("Missing check if access token is granted the required scopes")
-    hydraClient := hydra.NewHydraClient(aconf.HydraConfig)
-
-    log.WithFields(logrus.Fields{"token": accessToken.AccessToken}).Debug("Introspecting token")
-
-    introspectRequest := hydra.IntrospectRequest{
-      Token: accessToken.AccessToken,
-      Scope: strRequiredScopes, // This will make hydra check that all scopes are present else introspect.active will be false.
-    }
-    introspectResponse, err := hydra.IntrospectToken(aconf.HydraIntrospectUrl, hydraClient, introspectRequest)
-    if err != nil {
-      log.WithFields(logrus.Fields{"scope": strRequiredScopes}).Debug(err.Error())
-      c.AbortWithStatus(http.StatusInternalServerError)
-      return
-    }
-    log.Debug(introspectResponse)
-
-
-    var grantedScopes []string
-    var missingScopes []string
-
-    if introspectResponse.Active == true {
-
-      // Check scopes. (is done by hydra according to doc)
-      // https://www.ory.sh/docs/hydra/sdk/api#introspect-oauth2-tokens
-
-      // See #4 of QTNA
-      //log.WithFields(logrus.Fields{"fixme": 1, "qtna": 4}).Debug("Missing check if the user or client giving the grants in the access token authorized to use the scopes granted")
-      sub := introspectResponse.Sub
-      //client := introspectResponse.ClientId
-
-      publisherEntity := aap.Identity{ Id:config.GetString("id") }
-      requestorEntity := aap.Identity{ Id:sub }
-      ownerEntity := aap.Identity{ Id:sub }
-
-      session, tx, err := aap.BeginReadTx(env.Driver)
-      if err != nil {
-        log.Debug(err.Error())
-        c.AbortWithStatus(http.StatusInternalServerError)
-        return
-      }
-      defer tx.Close() // rolls back if not already committed/rolled back
-      defer session.Close()
-
-      for _, scope := range requiredScopes {
-
-        verdict, err := aap.JudgeEntity(tx, publisherEntity, requestorEntity, ownerEntity, aap.Scope{ Name:scope })
-        if err != nil {
-          log.WithFields(logrus.Fields{ "publisher":publisherEntity.Id, "requestor":requestorEntity.Id, "owner":ownerEntity.Id, "scope":scope }).Debug(err.Error())
-          c.AbortWithStatus(http.StatusInternalServerError)
-          return
-        }
-
-        if verdict.Granted == true {
-          grantedScopes = append(grantedScopes, scope)
-        } else {
-          missingScopes = append(missingScopes, scope)
-        }
-
-      }
-
-      if len(missingScopes) == 0 && len(grantedScopes) == len(requiredScopes) {
-        log.WithFields(logrus.Fields{"sub": introspectResponse.Sub, "scope": strRequiredScopes}).Debug("Authorized")
-        c.Set("sub", introspectResponse.Sub)
-        c.Next() // Authentication successful, continue.
-        return
-      }
-    }
-
-    // Deny by default
-    var strMissingScopes string = strings.Join(missingScopes, " ")
-    log.WithFields(logrus.Fields{"scope": strMissingScopes}).Debug("Missing Scopes");
-    c.AbortWithStatusJSON(http.StatusForbidden, JsonError{ErrorCode: ERROR_MISSING_REQUIRED_SCOPES, Error: fmt.Sprintf("Missing required scopes: %s. Hint: Some required scopes are missing, invalid or not granted", strMissingScopes)})
-    return
-
-  }
-  return gin.HandlerFunc(fn)
-}
-
-func RequestId() gin.HandlerFunc {
-  return func(c *gin.Context) {
-    // Check for incoming header, use it if exists
-    requestID := c.Request.Header.Get("X-Request-Id")
-
-    // Create request id with UUID4
-    if requestID == "" {
-      uuid4, _ := uuid.NewV4()
-      requestID = uuid4.String()
-    }
-
-    // Expose it for use in the application
-    c.Set("RequestId", requestID)
-
-    // Set X-Request-Id header
-    c.Writer.Header().Set("X-Request-Id", requestID)
-    c.Next()
-  }
 }
 
 func stringInSlice(a string, list []string) (bool) {
